set GLFW_CLIENT_LIBRARY-related definitions for the config-module
The GLFW_CLIENT_LIBRARY CMake option can take the following values:
`opengl`, `glesv1` or `glesv2`. It controls which dependency glfw should link
to and also which headers should be included.

In build-time these decisions are based on #defines in `glfw_config.h` but
that file not installed.

This commit exposes these definitions for the client of glfw using
target_compile_definitions() for the imported library.

The GLFW_INCLUDE_ES1/ES2 definition is used in glfw3.h to decide which
gl header to include.

The GLFW_USE_GLESV1/V2 definitions can be used by the client of glfw to
detect which OpenGL dialect to use.